DELTABOT: Windows MSVC build fixes + setup notes
Two MSVC-only warnings promoted to errors by /WX that the Linux-only CI
does not catch. GCC/Clang do not flag either.

- src/handlers/hints.h: match REAPER SDK's `class MediaTrack` /
  `class MediaItem` forward declarations (C4099).
- src/reaper/catalog.cpp:159: static_cast the int64_t result of
  db.scalar_int() to int (C4244) — matches the existing pattern
  on line 164 of the same file.

DELTABOT_README.md documents the local setup (vcpkg openssl, cmake
+ MSVC BuildTools, install-into-REAPER steps).

+++ src/reaper/catalog.cpp
@@ -156,7 +156,7 @@ void build(DB& db, const std::string& reaper_version) {
     db.query("INSERT OR REPLACE INTO meta(key, value) VALUES('catalog_version', ?1)",
              {std::to_string(k_catalog_version)});
 
-    int indexed = db.scalar_int("SELECT COUNT(*) FROM actions");
+    int indexed = static_cast<int>(db.scalar_int("SELECT COUNT(*) FROM actions"));
     Log::info("Action catalog: indexed " + std::to_string(indexed) + " actions");
 }
